Enable bitcoin_hashes v0.14.0
Hardware devices like the smallest binary possible, currently if one
builds with latest `rust-bitcoin` and `rust-bip39` they get two versions
of `bitcoin_hashes` in the dependency graph because we don't support the
latest `bitcoin_hashes`.

We use a range dependency for `bitcoin_hashes`. We can support the
latest version with no code changes. However using the latest version
causes the MSRV to increase.

Add the latest version of `bitcoin_hashes` to the range dependency.